+++ core/lib/moengage_flutter.dart
@@ -476,7 +476,7 @@ class MoEngageFlutter {
   }
 
   /// Configure MoEngage SDK Logs
-  /// @param [logLevel] LogLevel for SDK logs
+  /// @param [logLevel] - [LogLevel] for SDK logs
   /// @param [isEnabledForReleaseBuild] If true, logs will be printed for the Release build. By default the logs are disabled for the Release build.
   void configureLogs(LogLevel logLevel,
       {bool isEnabledForReleaseBuild = false}) {